chore: rename metadata.js types.js
